webapi/webapi-webspeech-w3c-tests/webspeech/speechrecognitionresultlist_length-manual.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu. Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset="utf-8">
<title>Web Speech API Test: SpeechRecognitionResultList length attribute</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dvcs.w3.org/hg/speech-api/raw-file/tip/speechapi.html">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<p>Speak 3 words in 1 second</p>
<div id="log"></div>
<script>

var t = async_test("Test if the length attribute indicates how many results are represented in the item array", { timeout: 5000});

t.step(function () {
  var speechReco = new webkitSpeechRecognition();
  speechReco.continuous = true;
  speechReco.onresult = function (evt) {
    var actual_long = 0;
    for (var i = evt.resultIndex; i < evt.results.length; i++) {
      actual_long++;
    }
    assert_equals(actual_long, evt.results.length);
    t.done();
  };
  speechReco.onerror = t.step_func(function (evt) {
    assert_unreached("SpeechRecognition get error " + evt.error);
    t.done();
  });
  speechReco.start();
  setTimeout(function () {
    speechReco.stop();
  }, 1000);
});

</script>
